Add support for showing less thumbnails than images
If the gallery contains many images (100s or 1000s) then showing that
many thumbnails is not practical. Make it possible to display only the
first few thumbnails, however when an image is opened, clicking to the
right should cycle through all images, not just the displayed
thumbnails.

To achieve this we introduce a new property maxThumbnails, to be used
as:

  <vue-picture-swipe :items="images" :maxThumbnails="8"></vue-picture-swipe>

Also, generating the items to supply to the underlying PhotoSwipe
library by parsing the thumbnails becomes unfeasible because less
thumbnails could be shown than there are images. So, generate the items
from the provided ":items" property, which is shorter and more
straight-forward anyway.

When closing a big image we animate it to the thumbnail that belongs to
it. Adjust that, so if an image is opened that does not have a
corresponding thumbnail, then animate to coordinates 0,0, width: 100.